+++ packages/coil-extension/src/background/services/BackgroundScript.ts
@@ -633,8 +633,11 @@ export class BackgroundScript {
   ) {
     const frame = getFrameSpec(sender)
     const { tabId, frameId } = frame
+    const { requestId } = request.data
 
-    this.tabStates.logLastMonetizationCommand(frame, 'start')
+    this.activeTabLogger.log(`startWM called with ${requestId}`, frame)
+    this.tabStates.logLastMonetizationCommand(frame, 'start', requestId)
+
     // This used to be sent from content script as a separate message
     this.mayMonetizeSite(sender, request.data.initiatingUrl)
 
@@ -645,7 +648,13 @@ export class BackgroundScript {
     const userBeforeReAuth = this.store.user
     let emittedPending = false
     const emitPending = () => {
-      this.sendSetMonetizationStateMessage(frame, 'pending')
+      // Set the requestId so that DocumentMonetization#setState will ignore
+      // this message if tags are added and removed extremely fast.
+      this.sendSetMonetizationStateMessage(
+        frame,
+        'pending',
+        request.data.requestId
+      )
       emittedPending = true
     }
 
@@ -664,7 +673,6 @@ export class BackgroundScript {
     }
 
     this.log('startWebMonetization, request', request)
-    const { requestId } = request.data
 
     this.log('loading token for monetization', requestId)
     const token = await this.auth.getTokenMaybeRefreshAndStoreState()
@@ -675,7 +683,11 @@ export class BackgroundScript {
         console.warn('startWebMonetization cancelled; no token')
       }
       this.activeTabLogger.log('startWebMonetization cancelled; no token')
-      this.sendSetMonetizationStateMessage(frame, 'stopped')
+      this.sendSetMonetizationStateMessage(
+        frame,
+        'stopped',
+        request.data.requestId
+      )
       setUnavailable('token')
       return false
     }
@@ -684,10 +696,37 @@ export class BackgroundScript {
         console.warn('startWebMonetization cancelled; no active subscription')
       }
       this.activeTabLogger.log(
-        'startWebMonetization cancelled; no active subscription'
+        'startWebMonetization cancelled; no active subscription',
+        frame
       )
-      this.sendSetMonetizationStateMessage(frame, 'stopped')
+      this.sendSetMonetizationStateMessage(
+        frame,
+        'stopped',
+        request.data.requestId
+      )
       setUnavailable('subscription')
+      return false
+    }
+
+    // Check that this startWebMonetization invocation is still valid before
+    // we go ahead. Any operation that we `await`d on could have potentially
+    // masked state changes. e.g. `getTokenMaybeRefreshAndStoreState`
+    // (which will update `whoami`) which takes longer than it does to switch
+    // out a monetization tag.
+    if (
+      this.tabStates.getFrameOrDefault(frame).lastMonetization.requestId !==
+      requestId
+    ) {
+      // The pending message (if sent) will have been ignored on the content
+      // script side in this case too, so there's no need to send a stop, as
+      // will already have been stopped (for that id).
+      // If we sent a stopped message, it would need to be tagged with
+      // requestId, and would only ever be ignored (
+      // due to the new monetization request implied in the if condition)
+      this.activeTabLogger.log(
+        `startWebMonetization aborted; stale requestId: ${requestId}`,
+        frame
+      )
       return false
     }
 
@@ -699,10 +738,16 @@ export class BackgroundScript {
       this.tabStates.getFrameOrDefault(frame).lastMonetization.command
     if (lastCommand !== 'start' && lastCommand !== 'pause') {
       this.log('startWebMonetization cancelled via', lastCommand)
+      this.activeTabLogger.log(
+        `startWebMonetization cancelled via ${lastCommand}`,
+        frame
+      )
       return false
     }
 
     this.log('starting stream', requestId)
+    // We need to start this stream, even if we've already received a pause.
+    // That way we can "resume" it later.
     this.assoc.setStreamId(frame, requestId)
     this.assoc.setFrame(requestId, { tabId, frameId })
     this.streams.beginStream(requestId, {
@@ -715,7 +760,7 @@ export class BackgroundScript {
     if (lastCommand === 'pause') {
       // TODO: why do we need the timeout here ?
       setTimeout(() => {
-        this.doPauseWebMonetization(frame)
+        this.doPauseWebMonetization(frame, requestId)
       }, 0)
     }
     return true
@@ -778,10 +823,10 @@ export class BackgroundScript {
     }
   }
 
-  private doPauseWebMonetization(frame: FrameSpec) {
-    this.tabStates.logLastMonetizationCommand(frame, 'pause')
-    const id = this.assoc.getStreamId(frame)
+  private doPauseWebMonetization(frame: FrameSpec, requestId?: string) {
+    const id = requestId ?? this.assoc.getStreamId(frame)
     if (id) {
+      this.tabStates.logLastMonetizationCommand(frame, 'pause', id)
       this.log('pausing stream', id)
       this.streams.pauseStream(id)
       this.sendSetMonetizationStateMessage(frame, 'stopped')
@@ -789,11 +834,10 @@ export class BackgroundScript {
     return true
   }
 
-  private doResumeWebMonetization(frame: FrameSpec) {
-    this.tabStates.logLastMonetizationCommand(frame, 'resume')
-
-    const id = this.assoc.getStreamId(frame)
+  private doResumeWebMonetization(frame: FrameSpec, requestId?: string) {
+    const id = requestId ?? this.assoc.getStreamId(frame)
     if (id) {
+      this.tabStates.logLastMonetizationCommand(frame, 'resume', id)
       this.log('resuming stream', id)
       // Always set "pending" transition state and emit event
       // We do this just for consistency/back-compat sake
@@ -812,7 +856,10 @@ export class BackgroundScript {
     if (this.tabStates.get(getTab(sender)).stickyState === 'sticky') {
       return
     }
-    return this.doPauseWebMonetization(getFrameSpec(sender))
+    return this.doPauseWebMonetization(
+      getFrameSpec(sender),
+      request.data.requestId
+    )
   }
 
   resumeWebMonetization(request: ResumeWebMonetization, sender: MessageSender) {
@@ -821,7 +868,10 @@ export class BackgroundScript {
     if (this.tabStates.get(getTab(sender)).playState === 'paused') {
       return
     }
-    return this.doResumeWebMonetization(getFrameSpec(sender))
+    return this.doResumeWebMonetization(
+      getFrameSpec(sender),
+      request.data.requestId
+    )
   }
 
   private handleStreamsAbortEvent() {
@@ -839,8 +889,10 @@ export class BackgroundScript {
   }
 
   private doStopWebMonetization(frame: FrameSpec) {
-    this.tabStates.logLastMonetizationCommand(frame, 'stop')
     const requestId = this.assoc.getStreamId(frame)
+    if (requestId) {
+      this.tabStates.logLastMonetizationCommand(frame, 'stop', requestId)
+    }
     const closed = this._closeStreams(frame.tabId, frame.frameId)
     // May be noop other side if stop monetization was initiated from
     // ContentScript