+++ packages/coil-oauth-scripts/src/bin/env.ts
@@ -30,9 +30,30 @@ const ClientOptions = class Options extends GraphQlClient.Options {
   extraHeaders = CF_HEADERS
 }
 
-export async function login() {
+export async function login(dbg?: typeof console.log, whoAmI?: boolean) {
   const client = new GraphQlClient(new ClientOptions())
+  if (dbg) {
+    dbg('logging in')
+  }
   const token = await client.login(COIL_USER, COIL_PASSWORD)
+  if (dbg) {
+    dbg('logged in')
+  }
+  if (dbg) {
+    dbg('refreshing token')
+  }
   const btpToken = await client.refreshBtpToken(token)
+  if (dbg) {
+    dbg('refreshed token')
+  }
+  if (whoAmI) {
+    if (dbg) {
+      dbg('getting whoAmI')
+    }
+    await client.whoAmI(token)
+    if (dbg) {
+      dbg('got whoAmI')
+    }
+  }
   return { client, token, btpToken }
 }