+++ packages/webmonetization-testing-utils/src/test.tsx
@@ -19,6 +19,7 @@ function makeLogger(logElSelector: string) {
 
 ;(async function main() {
   const log = makeLogger('#log')
+  log(`document.readyState ${document.readyState}`)
 
   const suite = new MonetizationImplTest(
     document as MonetizationExtendedDocument,
@@ -30,9 +31,9 @@ function makeLogger(logElSelector: string) {
     await suite.test()
     log('<strong>all tests passed!</strong>\n', true)
     log('\n<strong>all tests passed!</strong>', false)
-  } catch (e) {
+  } catch (e: unknown) {
     log('')
     log('Last Error:')
-    log(e.message)
+    log((e as Error).message)
   }
 })()